requirements.txt
# Основные зависимости
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
schedule==1.2.0

# База данных
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # Для PostgreSQL
alembic==1.13.1  # Для миграций

# NLP и ML
transformers==4.36.2
torch==2.1.1
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.18.0

# Анализ данных
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0

# API и сети
requests==2.31.0
aiohttp==3.9.1
httpx==0.26.0

# Утилиты
emoji==2.10.0
pytz==2023.3.post1
python-dateutil==2.8.2
pydantic==2.5.2

# Логирование
loguru==0.7.2

# Для деплоя
gunicorn==21.2.0